requirements.txt
numpy>=1.21.0
scipy>=1.7.0
sounddevice>=0.4.0
soundfile>=0.10.0
Pillow>=8.0.0
matplotlib>=3.3.0
